unix: reproducible directory order (scandir)
This commit replaces the readdir() loop with scandir. This means a well-
defined, reproducible sort order, with the downside that we will need to
store ALL directory entries in memory to sort them.

I am too lazy to add a switch to change how it is supposed to work. This
is only a proof-of-concept.

Co-Authored-By: Bernhard M. Wiedemann <bwiedemann suse.de>
(cherry picked from commit 5063c9ce12c58f2512c2d00080d863d14b5d31f7)